requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aioredis==2.0.1
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
exceptiongroup==1.2.0
frozenlist==1.4.1
html5lib==1.1
idna==3.6
iniconfig==2.0.0
loguru==0.7.2
magic-filter==1.0.12
multidict==6.0.5
packaging==24.0
pluggy==1.4.0
pydantic==2.5.3
pydantic_core==2.14.6
pytest==8.1.1
pytest-asyncio==0.23.6
python-dotenv==1.0.1
redis==5.0.3
requests==2.31.0
six==1.16.0
soupsieve==2.5
tomli==2.0.1
typing_extensions==4.11.0
urllib3==2.2.1
webencodings==0.5.1
yarl==1.9.4
